feat: Add normalize service

+++ internal/core/domain/index.go
@@ -1,5 +1,5 @@
 package domain
 
 type ComixIndex struct {
-	Index map[string][]int `json:"indexes"`
+	Index map[string][]uint64 `json:"indexes"`
 }
